+++ components/Page/Material/TypesCreate.vue
@@ -42,6 +42,7 @@
 
 <script lang="ts" setup>
 import { MaterialType } from "@prisma/client";
+
 import { ErrorCode } from "~~/utils/error";
 
 interface Props {
