+++ server/api/[...].ts
@@ -1,4 +1,4 @@
-import { ErrorCode, apiError } from "~~/utils/error";
+import { apiError, ErrorCode } from "~~/utils/error";
 
 export default defineEventHandler(() => {
   throw apiError(ErrorCode.NotFound, "Route not found.");
