+++ server/api/contacts/index.post.ts
@@ -1,5 +1,6 @@
 import { enums, object, optional, size, string } from "superstruct";
 import { ContactType } from "@prisma/client";
+
 import { contact } from "~/server/logic";
 import { readValidatedBody } from "~/server/utils";
 import { uuid } from "~/server/validation";
