+++ server/api/groups/memberships/[id]/index.patch.ts
@@ -1,7 +1,7 @@
 import { boolean, date, object, optional, size, string } from "superstruct";
 
-import { membership } from "~/server/logic";
 import { getValidatedRouterParam, readValidatedBody } from "~/server/utils";
+import { membership } from "~/server/logic";
 import { uuid } from "~~/server/validation";
 
 const body = object({
