+++ server/api/groups/memberships/index.post.ts
@@ -1,7 +1,8 @@
 import { boolean, date, object, optional, size, string } from "superstruct";
+
+import { membership } from "~/server/logic";
 import { readValidatedBody } from "~/server/utils";
 import { uuid } from "~/server/validation";
-import { membership } from "~/server/logic";
 
 const body = object({
   function: optional(size(string(), 2, 50)),
