+++ server/api/members/[id]/index.get.ts
@@ -1,5 +1,5 @@
-import { member } from "~/server/logic";
 import { getValidatedRouterParam } from "~~/server/utils";
+import { member } from "~/server/logic";
 import { uuid } from "~~/server/validation";
 
 export default defineEventHandler(async (event) => {
