+++ server/api/studies/index.post.ts
@@ -1,4 +1,5 @@
 import { date, object, optional, size, string } from "superstruct";
+
 import { readValidatedBody } from "~/server/utils";
 import { study } from "~/server/logic";
 import { uuid } from "~/server/validation";
