+++ server/api/studies/options/[id]/index.delete.ts
@@ -1,8 +1,9 @@
-import { study } from "~/server/logic";
 import { getValidatedRouterParam } from "~~/server/utils";
+import { study } from "~/server/logic";
 import { uuid } from "~~/server/validation";
 
 export default defineEventHandler(async (event) => {
   const id = await getValidatedRouterParam(event, "id", uuid());
-  return await study.removeOption(id);
+
+  return await study.option.remove(id);
 });
