+++ server/api/studies/options/index.get.ts
@@ -1,5 +1,5 @@
 import { study } from "~/server/logic";
 
 export default defineEventHandler(async () => {
-  return await study.getAllOptions();
+  return await study.option.getAll();
 });
