+++ server/api/studies/options/index.post.ts
@@ -1,7 +1,8 @@
 import { enums, object, size, string } from "superstruct";
 import { Institution, StudyLevel } from "@prisma/client";
-import { study } from "~/server/logic";
+
 import { readValidatedBody } from "~/server/utils";
+import { study } from "~/server/logic";
 
 const body = object({
   name: size(string(), 2, 40),
@@ -11,5 +12,5 @@ const body = object({
 
 export default defineEventHandler(async (event) => {
   const data = await readValidatedBody(event, body);
-  return await study.createOption(data);
+  return await study.option.create(data);
 });
