+++ server/logic/authentik/user.ts
@@ -1,7 +1,7 @@
-import { authentikFetch } from "./";
-
 import { apiError, ErrorCode } from "~/utils/error";
 import { member } from "~/server/logic";
+
+import { authentikFetch } from "./";
 
 const PATH = "users/proteus";
 const PAGE_SIZE = 100;
