+++ server/plugins/apiError.ts
@@ -1,4 +1,5 @@
 import { STATUS_CODES } from "http";
+
 import {
   PrismaClientInitializationError,
   PrismaClientKnownRequestError,
