+++ server/prisma/seed/index.ts
@@ -1,18 +1,18 @@
-import materialType from "./materialType";
+import group from "./group";
 import material from "./material";
+import materialType from "./materialType";
+import member from "./member";
 import quote from "./quote";
 import study from "./study";
-import group from "./group";
-import member from "./member";
 
 export const seed = async () => {
   try {
-    await materialType();
+    await group();
     await material();
+    await materialType();
+    await member();
     await quote();
     await study();
-    await group();
-    await member();
   } catch (e) {
     console.error("Seeding failed!");
     console.error(e);
